Extracting common test logic
