+++ spec/common/spec_helper.rb
@@ -1,0 +1,3 @@
+require_relative './query_runner_shared_examples'
+require_relative './test_runner_shared_examples'
+require_relative './integration_shared_examples'
