+++ spec/python2/spec_helper.rb
@@ -1,1 +1,2 @@
 require_relative '../../lib/python2_runner'
+require_relative '../common/spec_helper'
