+++ spec/python2/test_runner_spec.rb
@@ -4,26 +4,9 @@ describe Python2TestHook do
   before(:all) { reload_python2_runner! }
 
   let(:hook) { Python2TestHook.new }
-  let(:file) { hook.compile(request) }
-  let!(:result) { hook.run!(file) }
+  let!(:result) { hook.run!(hook.compile(request)) }
 
-  context 'passes when test pass' do
-    let(:request) { struct(content: '
-def foo():
-  return 4', test: '
-def test_true_is_true(self):
-  self.assertTrue(True)') }
-    it { expect(result[0]).to match_array [['True is true', :passed, '']] }
-  end
-
-  context 'fails when test fails' do
-    let(:request) { struct(content: '
-def foo():
-  return 4', test: '
-def test_true_is_false(self):
-  self.assertTrue(False)') }
-    it { expect(result[0]).to match_array [['True is false', :failed, 'AssertionError: False is not true']] }
-  end
+  it_behaves_like 'common python test hook'
 
   context 'fails when some test pass and other fail' do
     let(:request) { struct(content: '
@@ -41,28 +24,5 @@ def test_false_is_true(self):
                                               ['True is true', :passed, ''],
                                               ['False is true', :failed, 'AssertionError: False is not true'],
                                           ] }
-  end
-
-  context 'accepts full-defined tests' do
-    let(:request) { struct(content: '
-def foo():
-  return 4', test: '
-
-class MyTest(unittest.TestCase):
-  def test_true(self):
-    self.assertTrue(True)') }
-    it { expect(result[0]).to match_array [['True', :passed, '']] }
-  end
-
-  context 'accepts multiple tests' do
-    let(:request) { struct(content: '
-def foo():
-  return 4', test: '
-def test_false(self):
-  self.assertFalse(False)
-
-def test_true(self):
-  self.assertTrue(True)') }
-    it { expect(result[0]).to match_array [['False', :passed, ''], ['True', :passed, '']] }
   end
 end
