+++ spec/python3/integration_spec.rb
@@ -10,41 +10,9 @@ describe 'python3 integration test' do
   end
   after(:all) { Process.kill 'TERM', @pid }
 
-  it 'answers a valid hash when submission is ok' do
-    response = bridge.run_tests!(test: 'def test_foo_returns_true(self):
-    self.assertFalse(foo())',
-                                 extra: '',
-                                 content: "def foo():\n    return False\n",
-                                 expectations: [])
+  it_behaves_like 'common python runner'
 
-    expect(response).to eq(response_type: :structured,
-                           test_results: [{result: '', status: :passed, title: 'Foo returns true'}],
-                           status: :passed,
-                           feedback: '',
-                           expectation_results: [],
-                           result: '')
-  end
-
-  it 'answers a valid hash when submission is ok with a full-defined test' do
-    response = bridge.run_tests!(
-        test: '
-class TestFoo(unittest.TestCase):
-    def test_true(self):
-        self.assertFalse(foo())',
-        extra: '',
-        content: "def foo():\n    return False\n",
-        expectations: [])
-
-    expect(response).to eq(response_type: :structured,
-                           test_results: [{result: '', status: :passed, title: 'True'}],
-                           status: :passed,
-                           feedback: '',
-                           expectation_results: [],
-                           result: '')
-  end
-
-
-  it 'answers a valid hash when submission is not ok' do
+  it 'answers a valid hash when submission is not ok because of a global name' do
     response = bridge.
         run_tests!(test: '
 def test_foo_returns_false(self):
@@ -61,51 +29,6 @@ def test_foo_returns_false(self):
                            response_type: :structured)
   end
 
-  it 'answers a valid hash when submission has syntax errors' do
-    response = bridge.
-        run_tests!(test: '
-def test_foo_returns_false(self):
-    self.assertFalse(foo())',
-                   extra: '',
-                   content: 'no#COMPILA!"',
-                   expectations: [])
-
-    expect(response).to eq(status: :errored,
-                           expectation_results: [],
-                           feedback: '',
-                           test_results: [],
-                           result: "Traceback (most recent call last):\n" +
-                                   "  File \"solution.py\", line 5, in <module>\n" +
-                                   "    no#COMPILA!\"\n" +
-                                   "NameError: name 'no' is not defined\n",
-                           response_type: :unstructured)
-  end
-
-
-  it 'answers a valid hash when query is ok' do
-    response = bridge.run_query!(extra: 'x = 3',
-                                 content: 'y = 6',
-                                 query: 'x + y')
-    expect(response).to eq(status: :passed, result: "=> 9\n")
-  end
-
-  it 'answers a valid hash when query is not ok' do
-    response = bridge.
-        run_query!(extra: '',
-                   content: '',
-                   query: 'foo.bar()')
-    expect(response[:status]).to eq(:failed)
-  end
-
-  it 'status => aborted when using an infinite recursion' do
-    response = bridge.
-        run_query!(extra: '',
-                   content: "def foo():\n    foo()",
-                   query: 'foo()')
-
-    expect(response[:status]).to eq(:failed)
-    expect(response[:result]).to include('maximum recursion depth exceeded')
-  end
 
   it 'exposes python version' do
     expect(bridge.info['language']).to include('version' => '3.7.3')
