+++ spec/python3/spec_helper.rb
@@ -1,3 +1,3 @@
 require_relative '../../lib/python3_runner'
-
+require_relative '../common/spec_helper'
